backend/requirements.txt
flask
flask-cors
faiss-cpu
numpy
pillow
torch
transformers
python-multipart
